requirements.txt
image==1.5.32
imageio==2.4.1
imagesize==1.2.0
imutils==0.5.3
numpy==1.18.5
opencv-contrib-python==4.1.2.30
# opencv-python==4.1.2.30
Pillow==7.0.0
requests==2.23.0
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
sklearn==0.0
# skimage
sklearn-pandas==1.8.0
torchsummary==1.5.1
torchtext==0.3.1
#torch==1.6.0
#torchvision==0.7.0
# https://download.pytorch.org/whl/cpu/torch-1.1.0-cp37-cp37m-linux_x86_64.whl
# https://download.pytorch.org/whl/cpu/torchvision-0.3.0-cp37-cp37m-linux_x86_64.whl
urllib3==1.24.3